testing dynamic rendering of each method
